Non-transparent FST support by fst_manager logic
Adjusted the FST Manager's logic to support non-transparent FST
using the old (transparent) MUX interface and implementation.

Each peer has a list of connections (interface-MAC pairs), and a
peer can now be identified by any one of those pairs, instead of
by a single MAC address as before.
On new connection establishment, MB IEs of already established
connections are scanned in order to identify the peer to which that
new connection is relevant.

Change-Id: Iecd6f2f401a38671c6226c069872af44ccadfaff