fst-manager: Add support for working with a non-transparent FST AP
Upon connect, need to find the MAC address of the other band on
the remote device to be able to connect to it.
Currently we assume that both bands have the same MAC address.

The above assumption is not true for non-transparent FST mode.
To fix this we extract the MAC addresses of the peer device from MB IEs.

Change-Id: Ib6ad967672fd56137e14aba966a45fb9b4c4afe0
CRs-Fixed: 1056155